requirements.txt
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.0
flask-sqlalchemy==3.1.1
greenlet==3.0.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
urllib3==2.0.7
werkzeug==3.0.1
zipp==3.17.0
